refactor: samle personkontrakten i v2
Fjerner DTO-ene til det gamle POST /rest/api/testscenarios/v2/opprett-endepunktet, som slettes i den tilhørende serverendringen.

Flytter delte enumtyper til person.v2 og erstatter CountryCode med String slik at nv-i18n kan fjernes. Landkoder kan være ISO alpha-2 eller alpha-3 i overgangsperioden.
