--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/AaregDto.java
@@ -1,6 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-import java.util.List;
-
-public record AaregDto(List<ArbeidsforholdDto> arbeidsforhold) {
-}
