--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/Arbeidsgiver.java
@@ -1,15 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-
-import static com.fasterxml.jackson.annotation.JsonTypeInfo.Id.NAME;
-
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-
-@JsonTypeInfo(use = NAME)
-@JsonSubTypes({
-        @JsonSubTypes.Type(value = OrganisasjonDto.class, name = "organisasjon"),
-        @JsonSubTypes.Type(value = PrivatArbeidsgiver.class, name = "privateArbeidsgiver")
-})
-public interface Arbeidsgiver {
-}
