--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/ArenaDto.java
@@ -1,7 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-import java.util.List;
-
-public record ArenaDto(List<ArenaSakerDto> saker) {
-
-}
