--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/ArenaSakerDto.java
@@ -1,21 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-import java.util.List;
-
-public record ArenaSakerDto(YtelseTema tema, SakStatus status, List<ArenaVedtakDto> vedtak) {
-
-    public enum YtelseTema {
-        AAP,
-        DAG,
-        FA,
-        EF,
-        SP,
-        BS
-    }
-
-    public enum SakStatus {
-        AKTIV,
-        INAKT,
-        AVSLU
-    }
-}
