--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/ArenaVedtakDto.java
@@ -1,19 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-import java.time.LocalDate;
-import java.util.List;
-
-public record ArenaVedtakDto(LocalDate fom, LocalDate tom, VedtakStatus status, Integer dagsats, List<ArenaMeldekort> meldekort) {
-
-    public enum VedtakStatus {
-        IVERK,
-        OPPRE,
-        INNST,
-        REGIS,
-        MOTAT,
-        GODKJ,
-        AVSLU,
-        INAKT,
-        AKTIV
-    }
-}
