--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/FamilierelasjonModellDto.java
@@ -1,15 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-import java.util.UUID;
-
-public record FamilierelasjonModellDto(Relasjon relasjon, UUID relatertTilId) {
-
-    public enum Relasjon {
-        EKTE,
-        SAMBOER,
-        BARN,
-        FAR,
-        MOR,
-        MEDMOR
-    }
-}
