--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/InntektYtelseType.java
@@ -1,49 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-/*
- * Ytvalg fra kodeverkene Loennsbeskrivelse, Næringsinntektbeskrivelse, PensjonEllerTrygdBeskrivelse, YtelseFraOffentligeBeskrivelse
- */
-public enum InntektYtelseType  {
-
-    // Loennsbeskrivelse - Lønn
-    FASTLØNN,
-    FERIEPENGER,
-    KOMMUNAL_OMSORGSLØNN_MM, // Tilkommet for K9
-
-    // YtelseFraOffentligeBeskrivelse - Ytelse utbetalt til person som er arbeidstaker/frilanser/ytelsesmottaker
-    AAP,
-    DAGPENGER, DAGPENGER_FISKER_HYRE,
-    FORELDREPENGER,
-    SVANGERSKAPSPENGER,
-    SYKEPENGER, SYKEPENGER_FISKER_HYRE,
-    OMSORGSPENGER,
-    OPPLÆRINGSPENGER,
-    PLEIEPENGER,
-    OVERGANGSSTØNAD_ENSLIG,
-    VENTELØNN,
-
-    FERIEPENGER_FORELDREPENGER, FERIEPENGER_SVANGERSKAPSPENGER,
-    FERIEPENGER_OMSORGSPENGER, FERIEPENGER_OPPLÆRINGSPENGER, FERIEPENGER_PLEIEPENGER,
-    FERIEPENGER_SYKEPENGER, FERIEPENGER_SYKEPENGER_FISKER_HYRE,
-    FERIETILLEGG_DAGPENGER, FERIETILLEGG_DAGPENGER_FISKER_HYRE,
-
-    // PensjonEllerTrygdBeskrivelse - Annen ytelse utbetalt til person
-    KVALIFISERINGSSTØNAD,
-
-    // Næringsinntektbeskrivelse, Ytelse utbetalt til person som er næringsdrivende, fisker/lott, dagmamma eller jord/skogbruker
-    FORELDREPENGER_NÆRING, FORELDREPENGER_NÆRING_DAGMAMMA, FORELDREPENGER_NÆRING_FISKER, FORELDREPENGER_NÆRING_JORDBRUK,
-    SVANGERSKAPSPENGER_NÆRING,
-    SYKEPENGER_NÆRING, SYKEPENGER_NÆRING_DAGMAMMA, SYKEPENGER_NÆRING_FISKER, SYKEPENGER_NÆRING_JORDBRUK,
-    OMSORGSPENGER_NÆRING, OMSORGSPENGER_NÆRING_DAGMAMMA, OMSORGSPENGER_NÆRING_FISKER, OMSORGSPENGER_NÆRING_JORDBRUK,
-    OPPLÆRINGSPENGER_NÆRING,
-    PLEIEPENGER_NÆRING, PLEIEPENGER_NÆRING_DAGMAMMA, PLEIEPENGER_NÆRING_FISKER, PLEIEPENGER_NÆRING_JORDBRUK,
-    DAGPENGER_NÆRING,
-    DAGPENGER_NÆRING_FISKER,
-
-    // Annen ytelse utbetalt til person som er næringsdrivende
-    ANNET,
-    VEDERLAG, VEDERLAG_DAGMAMMA,
-    LOTT_KUN_TRYGDEAVGIFT,
-    KOMPENSASJON_FOR_TAPT_PERSONINNTEKT
-    ;
-}
