--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/Kjønn.java
@@ -1,6 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-public enum Kjønn {
-    M,
-    K
-}
