--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/MedlemskapDto.java
@@ -1,16 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-import java.time.LocalDate;
-
-import com.neovisionaries.i18n.CountryCode;
-
-public record MedlemskapDto(LocalDate fom,
-                            LocalDate tom,
-                            CountryCode land,
-                            DekningsType trygdedekning) {
-
-    public enum DekningsType {
-        IHT_AVTALE,
-        FULL,
-    }
-}
