--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/Orgnummer.java
@@ -1,6 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-import com.fasterxml.jackson.annotation.JsonValue;
-
-public record Orgnummer(@JsonValue String value) {
-}
