--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/PermisjonDto.java
@@ -1,9 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-import java.time.LocalDate;
-
-public record PermisjonDto(Integer stillingsprosent,
-                           LocalDate fomGyldighetsperiode,
-                           LocalDate tomGyldighetsperiode,
-                           Permisjonstype permisjonstype) {
-}
