--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/PersonstatusDto.java
@@ -1,10 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-import java.time.LocalDate;
-
-public record PersonstatusDto(Personstatuser personstatus, LocalDate fom, LocalDate tom) {
-
-      public enum Personstatuser {
-        ABNR, ADNR, BOSA, DØD, FOSV, FØDR, UFUL, UREG, UTAN, UTPE, UTVA;
-    }
-}
