--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/PesysDto.java
@@ -1,4 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-public record PesysDto(Boolean harUføretrygd) {
-}
