--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/PrivatArbeidsgiver.java
@@ -1,6 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-import java.util.UUID;
-
-public record PrivatArbeidsgiver(UUID uuid) implements Arbeidsgiver {
-}
