--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/SigrunDto.java
@@ -1,8 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-import java.util.List;
-
-public record SigrunDto(List<InntektsårDto> inntektår) {
-    public record InntektsårDto(Integer år, Integer beløp) {
-    }
-}
