--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/SivilstandDto.java
@@ -1,21 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-import java.time.LocalDate;
-
-public record SivilstandDto(Sivilstander sivilstand, LocalDate fom, LocalDate tom){
-
-    public enum Sivilstander {
-        ENKE,
-        GIFT,
-        GJPA,
-        GLAD,
-        REPA,
-        SAMB,
-        SEPA,
-        SEPR,
-        SKIL,
-        SKPA,
-        UGIF;
-    }
-
-}
