--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/Språk.java
@@ -1,7 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-public enum Språk {
-    NB,
-    NN,
-    EN
-}
