--- kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/StatsborgerskapDto.java
@@ -1,6 +1,0 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
-
-import com.neovisionaries.i18n.CountryCode;
-
-public record StatsborgerskapDto(CountryCode land) {
-}
