+++ kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/v2/AdresseDto.java
@@ -2,10 +2,8 @@ package no.nav.foreldrepenger.vtp.kontrakter.person.v2;
 
 import java.time.LocalDate;
 
-import com.neovisionaries.i18n.CountryCode;
-
 public record AdresseDto(AdresseType adresseType,
-                         CountryCode land,
+                         String land,
                          String matrikkelId,
                          LocalDate fom,
                          LocalDate tom) {
