+++ kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/v2/Arbeidsforholdstype.java
@@ -1,4 +1,4 @@
-package no.nav.foreldrepenger.vtp.kontrakter.person;
+package no.nav.foreldrepenger.vtp.kontrakter.person.v2;
 
 public enum Arbeidsforholdstype {
     ORDINÆRT_ARBEIDSFORHOLD,
