+++ kontrakter/src/main/java/no/nav/foreldrepenger/vtp/kontrakter/person/v2/Språk.java
@@ -1,4 +1,7 @@
 package no.nav.foreldrepenger.vtp.kontrakter.person.v2;
 
-public record StatsborgerskapDto(String land) {
+public enum Språk {
+    NB,
+    NN,
+    EN
 }
